+++ CHANGELOG.md
@@ -15,11 +15,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
--
+- New property and methods overloads to `ex.Animation`
+  * `ex.Animation.currentFrameTimeLeft` will return the current time in milliseconds left in the current
+  * `ex.Animation.goToFrame(frameNumber: number, duration?: number)` now accepts an optional duration for the target frame
 
 ### Fixed
 
--
+- `ex.Animation.reset()` did not properly reset all internal state
 
 ### Updates